+++ src/utils.cpp
@@ -1,4 +1,4 @@
-#include "utils.h"
+#include "utils.hpp"
 #include <chrono>
 #include <random>
 #if defined(_WIN32)
